BFN Repacking
As of this commit, bfntoolkit can successfully repack folders into BFN files, provided you put bti versions of the font textures into the folder beforehand.  bfntoolkit will have to convert the textures itself in the future due to the fact that there are a number of differences between textures in a BFN file and individual BTI files.